chore: ensure effect destruction of deriveds is consistent
