+++ packages/svelte/src/internal/client/runtime.js
@@ -2,7 +2,10 @@
 import { DEV } from 'esm-env';
 import { define_property, get_descriptors, get_prototype_of } from '../shared/utils.js';
 import {
+	destroy_block_effect_children,
 	destroy_effect,
+	destroy_effect_children,
+	destroy_effect_deriveds,
 	effect,
 	execute_effect_teardown,
 	unlink_effect
@@ -404,48 +407,6 @@ export function remove_reactions(signal, start_index) {
 }
 
 /**
- * @param {Effect} signal
- * @param {boolean} remove_dom
- * @returns {void}
- */
-export function destroy_effect_children(signal, remove_dom = false) {
-	var deriveds = signal.deriveds;
-
-	if (deriveds !== null) {
-		signal.deriveds = null;
-
-		for (var i = 0; i < deriveds.length; i += 1) {
-			destroy_derived(deriveds[i]);
-		}
-	}
-
-	var effect = signal.first;
-	signal.first = signal.last = null;
-
-	while (effect !== null) {
-		var next = effect.next;
-		destroy_effect(effect, remove_dom);
-		effect = next;
-	}
-}
-
-/**
- * @param {Effect} signal
- * @returns {void}
- */
-export function destroy_block_effect_children(signal) {
-	var effect = signal.first;
-
-	while (effect !== null) {
-		var next = effect.next;
-		if ((effect.f & BRANCH_EFFECT) === 0) {
-			destroy_effect(effect);
-		}
-		effect = next;
-	}
-}
-
-/**
  * @param {Effect} effect
  * @returns {void}
  */
@@ -470,6 +431,7 @@ export function update_effect(effect) {
 	}
 
 	try {
+		destroy_effect_deriveds(effect);
 		if ((flags & BLOCK_EFFECT) !== 0) {
 			destroy_block_effect_children(effect);
 		} else {
